separacion de los componentes y de sus pruebas en archivos independientes
