+++ Implementacion/agendador/src/FormularioEvento.js
@@ -1,0 +1,68 @@
+/**
+  Componente que representa el formulario para la creacion de eventos. 
+  Recibe como propiedad el callback onSubmit(evento) que se ejecutara cuando el evento halla sido terminado de crear.
+  @Autor Diego Felipe Orozco Penagos
+*/
+import React from 'react';
+import {
+  StyleSheet,
+  View,
+  Text,
+  Button,
+  TextInput
+} from 'react-native';
+
+import generalStyles from '../App';
+import {timeToString} from './dateFunctions';
+import DateTimePicker from '@react-native-community/datetimepicker';
+
+export default function FormularioEvento(props){
+  const [error, indicarError] = React.useState(null);
+  const [name, setName] = React.useState('');
+  const [horaI, setHoraI] = React.useState({hora: new Date(), show: false});
+  const [horaF, setHoraF] = React.useState({hora: new Date(), show: false});
+
+  function submitEvento(){
+    //validaciones de campos
+    if (name.trim() === ''){
+      indicarError(<Text style={StyleSheet.flatten([generalStyles.visualViews, generalStyles.errors])}>debe indicar el nombre del evento</Text>);
+    } else if (horaF.hora - horaI.hora <= 0){
+      indicarError(<Text style={StyleSheet.flatten([generalStyles.visualViews, generalStyles.errors])}>La hora de inicio debe ser menor ni igual a la hora de fin</Text>)
+    } else {
+      props.onSubmit({
+        key: name,
+        name: name,
+        horaI: timeToString(horaI.hora),
+        horaF: timeToString(horaF.hora)
+      });
+    }
+  }
+  return(
+    <View>
+      {error}
+      <Text style={generalStyles.visualViews}>nombre*</Text>
+      <TextInput style={StyleSheet.flatten([generalStyles.visualViews, generalStyles.borderBlue])} onChangeText={name => {setName(name)}} testID='nombre'/>
+      <Button style={generalStyles.visualViews} title={'Hora de inicio: ' + timeToString(horaI.hora)} onPress={()=>setHoraI(hora=>({...hora, show: true}))} testID='horaInicio'/>
+      <View style={generalStyles.separador}></View>
+      <Button style={generalStyles.visualViews} title={'Hora de fin: ' + timeToString(horaF.hora)} onPress={()=>setHoraF(hora=>({...hora, show:true}))} testID='horaFin'/>
+      <View style={generalStyles.separador}></View>
+      <Button style={{marginTop: 10, marginBottom: 10}} title='Terminado' onPress={()=>submitEvento()} testID='submit'/>
+      {horaI.show && <DateTimePicker 
+                          value={horaI.hora | new Date()}
+                          mode='time'
+                          is24Hour={true}
+                          display='default'
+                          onChange={(event, time=new Date())=>{setHoraI({hora: time, show: false})}} 
+                          testID='calendarI'/>
+      }
+      {horaF.show && <DateTimePicker 
+                          value={horaF.hora | new Date()}
+                          mode='time'
+                          is24Hour={true}
+                          display='default'
+                          onChange={(event, time=new Date())=>{setHoraF({hora: time, show: false})}} 
+                          testID='calendarF'/>
+      }
+    </View>
+  );
+}
